+++ CommerceFlow.Persistence/Abstractions/IUsersRepository.cs
@@ -4,10 +4,10 @@ namespace CommerceFlow.Persistence.Abstractions
 {
     public interface IUsersRepository
     {
-        Task<ulong> AddAsync(User user);
-        Task<int> DeleteAsync(ulong id);
+        Task<int> AddAsync(User user);
+        Task<int> DeleteAsync(int id);
         Task<List<User>> GetAllAsync();
-        Task<User> GetAsync(ulong id);
+        Task<User> GetAsync(int id);
         Task<int> UpdateAsync(User user);
     }
 }
